# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: ChiefMKT AI-powered digital marketing platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects/websites
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{projectId}/tracking-snippet:
    get:
      operationId: getTrackingSnippet
      tags: [tracking]
      summary: Get JavaScript tracking snippet for project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tracking snippet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingSnippet"
...
  /funnels:
    get:
      operationId: listFunnels
      tags: [funnels]
      summary: List funnels for a project
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Funnels list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Funnel"
    post:
      operationId: createFunnel
      tags: [funnels]
      summary: Create a new funnel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFunnelRequest"
      responses:
        "201":
          description: Funnel created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Funnel"
  /funnels/{funnelId}/data:
    get:
      operationId: getFunnelData
      tags: [funnels]
      summary: Get funnel analytics data
      parameters:
        - name: funnelId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Funnel data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FunnelData"
  /ab-tests:
    get:
      operationId: listAbTests
      tags: [ab-testing]
      summary: List A/B tests
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: A/B tests list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AbTest"
    post:
      operationId: createAbTest
      tags: [ab-testing]
      summary: Create an A/B test
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAbTestRequest"
      responses:
        "201":
          description: A/B test created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AbTest"
...
  /email/campaigns:
    get:
      operationId: listEmailCampaigns
      tags: [email]
      summary: List email campaigns
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Email campaigns
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EmailCampaign"
    post:
      operationId: createEmailCampaign
      tags: [email]
      summary: Create email campaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEmailCampaignRequest"
      responses:
        "201":
          description: Email campaign created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmailCampaign"
  /social/posts:
    get:
      operationId: listSocialPosts
      tags: [social]
      summary: List scheduled social posts
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Social posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SocialPost"
    post:
      operationId: createSocialPost
      tags: [social]
      summary: Create/schedule a social post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSocialPostRequest"
      responses:
        "201":
          description: Social post created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialPost"
  /competitors:
    get:
      operationId: listCompetitors
      tags: [competitors]
      summary: List tracked competitors
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Competitors list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Competitor"
    post:
      operationId: addCompetitor
      tags: [competitors]
      summary: Add a competitor to track
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddCompetitorRequest"
      responses:
        "201":
          description: Competitor added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Competitor"
  /chat-widget/settings:
    get:
      operationId: getChatWidgetSettings
      tags: [chat-widget]
      summary: Get chat widget settings
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat widget settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatWidgetSettings"
    put:
      operationId: updateChatWidgetSettings
      tags: [chat-widget]
      summary: Update chat widget settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatWidgetSettings"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatWidgetSettings"
  /chat-widget/conversations:
    get:
      operationId: listChatConversations
      tags: [chat-widget]
      summary: List chat conversations
      parameters:
        - name: projectId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Chat conversations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatConversation"
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        trackingId:
          type: string
        industry:
          type: string
        createdAt:
          type: string
          format: date-time
        isActive:
          type: boolean
      required:
        - id
        - name
        - url
        - trackingId
        - isActive
        - createdAt
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        industry:
          type: string
      required:
        - name
        - url
    TrackingSnippet:
      type: object
      properties:
        projectId:
          type: integer
        trackingId:
          type: string
        snippet:
          type: string
      required:
        - projectId
        - trackingId
        - snippet
...
    Funnel:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        projectId:
          type: integer
        steps:
          type: array
          items:
            $ref: "#/components/schemas/FunnelStep"
        createdAt:
          type: string
          format: date-time
        isActive:
          type: boolean
      required:
        - id
        - name
        - projectId
        - steps
        - createdAt
        - isActive
    FunnelStep:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        order:
          type: integer
      required:
        - id
        - name
        - url
        - order
    CreateFunnelRequest:
      type: object
      properties:
        name:
          type: string
        projectId:
          type: integer
        steps:
          type: array
          items:
            type: object
            properties:
              name:
                type: string
              url:
                type: string
              order:
                type: integer
            required:
              - name
              - url
              - order
      required:
        - name
        - projectId
        - steps
    FunnelData:
      type: object
      properties:
        funnelId:
          type: integer
        steps:
          type: array
          items:
            $ref: "#/components/schemas/FunnelStepData"
        overallConversionRate:
          type: number
        totalEntries:
          type: integer
      required:
        - funnelId
        - steps
        - overallConversionRate
        - totalEntries
    FunnelStepData:
      type: object
      properties:
        stepId:
          type: integer
        name:
          type: string
        visitors:
          type: integer
        dropoffRate:
          type: number
        conversionRate:
          type: number
      required:
        - stepId
        - name
        - visitors
        - dropoffRate
        - conversionRate
    AbTest:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        projectId:
          type: integer
        status:
          type: string
          enum: [draft, running, paused, completed]
        control:
          $ref: "#/components/schemas/AbTestVariant"
        variant:
          $ref: "#/components/schemas/AbTestVariant"
        winner:
          type: string
          nullable: true
        startedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        confidence:
          type: number
      required:
        - id
        - name
        - projectId
        - status
        - control
        - variant
        - createdAt
        - confidence
    AbTestVariant:
      type: object
      properties:
        name:
          type: string
        url:
          type: string
        visitors:
          type: integer
        conversions:
          type: integer
        conversionRate:
          type: number
      required:
        - name
        - url
        - visitors
        - conversions
        - conversionRate
    CreateAbTestRequest:
      type: object
      properties:
        name:
          type: string
        projectId:
          type: integer
        controlUrl:
          type: string
        variantUrl:
          type: string
        goal:
          type: string
      required:
        - name
        - projectId
        - controlUrl
        - variantUrl
        - goal
...
    EmailCampaign:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        subject:
          type: string
        status:
          type: string
          enum: [draft, scheduled, sent, paused]
        recipients:
          type: integer
        opens:
          type: integer
        clicks:
          type: integer
        openRate:
          type: number
        clickRate:
          type: number
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        sentAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - subject
        - status
        - recipients
        - opens
        - clicks
        - openRate
        - clickRate
        - createdAt
    CreateEmailCampaignRequest:
      type: object
      properties:
        name:
          type: string
        subject:
          type: string
        body:
          type: string
        projectId:
          type: integer
        scheduledAt:
          type: string
          format: date-time
          nullable: true
      required:
        - name
        - subject
        - body
        - projectId
    SocialPost:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        platforms:
          type: array
          items:
            type: string
            enum: [twitter, linkedin, facebook, instagram]
        status:
          type: string
          enum: [draft, scheduled, published, failed]
        scheduledAt:
          type: string
          format: date-time
          nullable: true
        publishedAt:
          type: string
          format: date-time
          nullable: true
        likes:
          type: integer
        shares:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - content
        - platforms
        - status
        - likes
        - shares
        - createdAt
    CreateSocialPostRequest:
      type: object
      properties:
        content:
          type: string
        platforms:
          type: array
          items:
            type: string
        projectId:
          type: integer
        scheduledAt:
          type: string
          format: date-time
          nullable: true
      required:
        - content
        - platforms
        - projectId
    Competitor:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        url:
          type: string
        projectId:
          type: integer
        metrics:
          type: object
          properties:
            domainAuthority:
              type: integer
            backlinks:
              type: integer
            organicKeywords:
              type: integer
            estimatedTraffic:
              type: integer
          required:
            - domainAuthority
            - backlinks
            - organicKeywords
            - estimatedTraffic
        lastUpdated:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - url
        - projectId
        - metrics
        - createdAt
        - lastUpdated
    AddCompetitorRequest:
      type: object
      properties:
        url:
          type: string
        projectId:
          type: integer
      required:
        - url
        - projectId
    ChatWidgetSettings:
      type: object
      properties:
        projectId:
          type: integer
        isEnabled:
          type: boolean
        welcomeMessage:
          type: string
        botName:
          type: string
        primaryColor:
          type: string
        position:
          type: string
          enum: [bottom-right, bottom-left]
        qualifyLeads:
          type: boolean
        captureEmail:
          type: boolean
      required:
        - projectId
        - isEnabled
        - welcomeMessage
        - botName
        - primaryColor
        - position
        - qualifyLeads
        - captureEmail
    ChatConversation:
      type: object
      properties:
        id:
          type: integer
        visitorId:
          type: string
        email:
          type: string
          nullable: true
        messagesCount:
          type: integer
        isQualified:
          type: boolean
        startedAt:
          type: string
          format: date-time
        lastMessageAt:
          type: string
          format: date-time
      required:
        - id
        - visitorId
        - messagesCount
        - isQualified
        - startedAt
        - lastMessageAt
...